+++ packages/terra-framework-docs/src/terra-dev-site/doc/embedded-content-consumer/providers/BasicProviderWithScrolling.provider.jsx
@@ -1,15 +1,9 @@
 import React from 'react';
-import { Provider } from 'xfc';
 import classNames from 'classnames/bind';
 import ProviderTestTemplate from 'terra-embedded-content-consumer/lib/EmbeddedContentProviderTestTemplate';
 import styles from './EmbeddedContentConsumerCommon.module.scss';
 
 const cx = classNames.bind(styles);
-
-Provider.init({
-  acls: ['*'],
-  secret: () => (Promise.resolve('Success')),
-});
 
 const BasicProviderWithScrolling = () => (
   <ProviderTestTemplate>